Remove unsafe from EqualStartingCharacterCount (#121102)
I wasn't able to detect any difference for `Path.GetRelativePath` and
the codegen is fairly close.
if we want extra perf, we can vectorize this routine trivially for
Linux/macOS (where paths are case-sensitives).

Given this API allocates a new string, it probably isn't supposed to be
used on a hot path.

---------

Co-authored-by: xtqqczze <45661989+xtqqczze@users.noreply.github.com>