Fix CborWriter.Reset.
The first issue is that Reset did not correctly set _definiteLength back after a reset. If allowMultipleRootLevelValues from the constructor is false, the default, it should be initialized to 1. However Reset set it as null, which behaves as if allowMultipleRootLevelValues is true. This sets _definiteLength back to the same way the constructor does.

The second is that _currentIndefiniteLengthStringRanges is not cleared during Reset. This means if a Reset occurred while in the middle of writing an indefinite length value and convertIndefiniteLengthEncodings is true, the next time an encode tried to re-encode segments as definite encoding where the list indicated would be pointing to incorrect locations.