Fix handling exceptions from native code when interpreted
There were two issues. One is that the SfiNextWorker was incorrectly
advancing the stack frame iterator in case the interpreter was called
from CallDescrWorker in some cases.
The second was that the call to MethodDesc::GetMethodDescOfVirtualizedCode
that is made by the InterpExecMethod can end up running managed code and
throw managed exception. So it needs to have PAL_TRY/PAL_EXCEPT around
the invocation, since the main loop of the interpreter can only catch
and process C++ exceptions.