Add EventPipe EventSource serialization testing (#120887)
We've had EventSource serialization tests for a long time but EventPipe
never got added to it. Unforetunately not testing EventPipe in this way
previously meant that quite a few discrepancies between EventPipe and
ETW went unnoticed. As a first step to documenting and/or resolving the
problems I wanted to get EventPipe running in the same set of tests. I
also did several other small fixes and improvements while doing the
work.

- EventPipeListener added as a new 3rd target for EventSource (in
addition to EventListener and ETW which are there now)
- Modified the Listener test abstraction so that EventPipe could also
work
- Removed assumption that providers can changed after the listener
starts running
- Added capability flags to indicate whether listeners supported certain
kinds of serialization behavior
- Worked around incomplete implementation of IDictionary for payloads in
TraceEvent
- Leveraged xunit theory attribute to avoid needing different top-level
APIs for every listener configuration
- Fixed EventTestHarness assert which had actual and expected values
swapped
- TestsWrite: Fixed mistaken call to Array.Equals with Assert.Equal so
the test actually validates what it was supposed to
- Removed ActiveIssue 88305 attribute which was already closed in the
past.
- Moved some test cases from TestsWriteEvent.Etw.cs to
TestsWriteEvent.cs so that they can run for other listeners when
supported.