Capture PDBs in `--make-repro-path` (#121163)
Repro for dotnet/roslyn#80952 was somewhat harder to obtain because we
don't pack PDBs into the repro package (they are not command line
arguments and the compiler looks for them next to the assembly).

This is a simple change to just look for .pdb files next to the files
we're packing.

Cc @dotnet/ilc-contrib