Fix interpreter tailcall to pinvoke
The interpreter compiler ignores tail call prefix for tail calls (and
jmp) to pinvokes that are not marshalled. But it is not injecting "ret"
IR opcode after the call returns, which results in BADCODE reported by
the compiler.

There are several coreclr tests in the JIT/jit64/mcc/interop that fail
due to that.

This change adds injecting proper form of "ret" IR opcode after such
calls.