[wasm][coreclr] Fix native calls to instance methods with 0 parameters (#121312)
The `nSizeOfArgStack` was calculated wrong, because the size of This
argument was not included. For cases with more than zero parameters it
is included in first arg offset.

It is different on wasm, because we don't use register for This argument
as all the other architectures.

It fixes the problem in
https://github.com/dotnet/runtime/issues/121107 where the
`GetExceptionMessage` call ended without the `OBJECTREF throwable` on
the stack, resulting in random pointer instead.

---------

Co-authored-by: Jan Kotas <jkotas@microsoft.com>