Fix generic argument access in filter funclets (#121332)
In funclets of generic methods, access to the generic argument may be
needed. It doesn't work correctly, because the filter funclet has its
own locals and various IR instructions that use that argument cannot
access it.

This change adds copying of that argument from the parent frame to the
funclet frame at the same offset so that the access to it just works.