[mono][sgen] Make color visible to client permanent (#121247)
A color (SCC) that isn't containing any bridge objects is made visible
to client if xrefs_in * xrefs_out is greater than 60. Later on in bridge
processing, we need to build the callback to pass for .net android.
During this stage, we reduce from the full set of SCCs to SCCs that
should be visible to client (containing bridge objects or satisfying the
above condition). If an SCC has an xref to a color that is not visible
to client, we need to do a recursive traversal to find all neighbors
that are visible to client. The problem is that this process can end up
making an SCC no longer visible to client, leading to inconsistencies in
the computation. Consider a color(C1) that has a neighbor that is not
visible to client(C2). In this final stage, we compute the neighbors of
C1 by traversing recursively through the neighbors of C2. If C2 ends up
pointing to colors that were already neighbors of C1, then, following
this computation, C1 would end up with fewer xrefs_out, making the color
no longer visible to client. This make future checks incorrect,
resulting in building incorrect graph for client.

This scenario seems rare in practice, we should have gotten way more
reports otherwise. We fix this by pinning the visible_to_client property
for a color once it first satisfies it, so it will no longer matter how
many actual xrefs the color has.

Fixes assertions like:
```
* Assertion at /home/vbrezae/Xamarin/repos/runtime/src/mono/mono/metadata/sgen-tarjan-bridge.c:1151, condition `color_visible_to_client (cd)' not met
```