First part of async support in native AOT (#121295)
Following works with runtime async enabled in Roslyn on native AOT:

```csharp
public class Async2Void
{
    public static void Main()
    {
        var t = AsyncTestEntryPoint(123, 456);
        t.Wait();
        Console.WriteLine(t.Result);
    }

    private static async Task<int> AsyncTestEntryPoint(int x, int y)
    {
        int result = await OtherAsync(x, y);
        return result;
    }

    private static async Task<int> OtherAsync(int x, int y)
    {
        Console.WriteLine(x);
        Console.WriteLine(y);
        return x + y;
    }
}
```

The implementation strategy is as follows:

* In the compiler, whenever someone holds an `EcmaMethod` of an asyncv2
method, it means they're holding the `RuntimeAsync` flavor of the method
in CoreCLR VM parlance. The return value is a `Task`,
`MethodDesc.IsAsync` is true, `MethodSignature.IsAsyncCallConv` is
false. The IL is a thunk to the `AsyncVariantImplMethod` flavor.
* In the compiler, we have a new `MethodDesc` descendant:
`AsyncVariantImplMethod`. This is the actual method with async calling
convention that corresponds to some `EcmaMethod` with `IsAsync`==true.
For this one: the return value is unwrapped from `Task`,
`MethodDesc.IsAsync` is true, `MethodSignature.IsAsyncCallConv` is true.
The IL is the actual IL that corresponds to the wrapped `EcmaMethod` on
disk.

Cc @dotnet/ilc-contrib