Fix unhandled exception test for interpreter
The test was accepting only 0xc0000005 as an exit code for
NullReferenceException. But in the interpreter case, the exception is
thrown using COMPlusThrow(kNullReferenceException) and gets the error
code of regular managed exception, the 0xE0434352.

This change updates the test to accept that one too and adds
verification that the exception message contains the
System.NullReferenceException to make sure the tested app didn't exit
due to some other exception.