[RuntimeAsync] Fix for configured ValueTask sources (#120704)
Fixes: #120709

There is a scenario where `ValueTask` decides on whether to run
continuations on the scheduling context or not.

It happens in a case when an `await` calls into a `ValueTask` that wraps
a `ValueTaskSource`.
In such case we may have a nontrivial context provided to the
`ValueTaskSource` at the time of the await, but `ValueTaskSource` may
choose to ignore the context when running callbacks.

A typical test sensitive to this looks like:


https://github.com/dotnet/runtime/blob/ce717acc70b4f165be5ab4de5053ef604287fd9f/src/libraries/System.IO.Pipelines/tests/SchedulerFacts.cs#L48-L89

The part
```cs
                    ReadResult result = await pipe.Reader.ReadAsync();
```

should behave as
```cs
                    ReadResult result = await pipe.Reader.ReadAsync().ConfigureAwait(false);
```

Otherwise we end up posting the `doRead` continuation to the tracking
`CustomSynchronizationContext` which only tracks Posts and does not run
them, thus the test makes no progress.

---------

Co-authored-by: Jakob Botsch Nielsen <Jakob.botsch.nielsen@gmail.com>