Make `EnumAllVirtualSlots` return slots for interfaces
Calling `EnumAllVirtualSlots` on classes returns all `newslot` methods that exist in the class hierarchy. Calling this on interfaces produced an empty enumeration.

For runtime-async, it would be convenient if this returned the logical slots on interfaces. While interfaces don't have physical slots, they still have logical slots, so it's not a big stretch.

Doing this now simplifies things in some places, and requires an extra `if` check in other places. But it will be really convenient for runtime-async.