Async continuations for native AOT (#121398)
The same program as in #121295 still works, but we can newly report async helpers as Intrinsic/Async.

This implements things around suspension/resumption. Most of this change is around handling continuation types.

Continuation types are synthetic types (created in the compiler) that derive from `Continuation` in the CoreLib. The JIT requests these based on the shape of locals it needs to preserve. What we get from the JIT in CorInfoImpl is size of the type and a pointer map (1011001 - non-zero means "GC pointer is here"). What we need to generate is a `MethodTable` for a type that derives from `Continuation` and has the specified GC layout after fields inherited from `Continuation`.

We already have a similar thing in the compiler ("`MethodTable`" we use for GC statics), so we're reusing the same approach. The `MethodTable` is pretty restricted and not the whole `MethodTable` API surface on it is available at runtime. However, it implements enough that the GC can operate on it.

The two new types are `AsyncContinuationType`. This is a type system `MetadataType` that represents the continuation pointer map. Then we have `AsyncContinuationEETypeNode` which is the thing that gets emitted into the output file - the `MethodTable` itself.

The resumption stub is just a stubbed out TODO for someone else to look into for now.