[RuntimeAsync] Do not hold on to the last used Continuation in the runtime async tasks. (#121460)
Fixes: #121459

The wrapped continuation is expected to be used only once - when the
corresponding async method is resumed. We can and should null-out the
continuation while fetching it in the dispatcher. There is no need to
keep it referenced from the task once fetched as it may cause "leaks"
observable via WeakReference/Finalization and the like.

The close analog of this behavior in the async1 is the C# state machine
zeroing out pointer-containing fields in the display struct upon
completing or faulting the task.


https://github.com/dotnet/roslyn/blob/f2df0ad6d3ef30ed033a794e7150a5600be50a95/src/Compilers/CSharp/Portable/Lowering/AsyncRewriter/AsyncMethodToStateMachineRewriter.cs#L186


https://github.com/dotnet/roslyn/blob/f2df0ad6d3ef30ed033a794e7150a5600be50a95/src/Compilers/CSharp/Portable/Lowering/AsyncRewriter/AsyncMethodToStateMachineRewriter.cs#L224