[release/10.0] Add warning for annotated extension property (#120005) (#120118)
Backport of https://github.com/dotnet/runtime/pull/120005 to 10.0.

## Customer Impact

- [ ] Customer reported
- [x] Found internally

This fixes an issue where `DynamicallyAccessedMembers` on extension
properties (new in .NET 10) is silently ignored. The existing behavior
violates user expectations because the attribute has no effect. With the
fix, this will be a warning.

## Regression

- [ ] Yes
- [x] No

## Testing

Updated tests to validate the new behavior. This behavior was known and
tested in the initial change that introduced support for extension
members. and we decided after the fact that this should produce a
warning.

## Risk

Low. This fix is added for completeness in a new .NET 10 feature
(trimming support for C# 14 extensions) and does not impact code that
doesn't use C# 14 extensions.

---------

Co-authored-by: Andy Gocke <angocke@microsoft.com>