[release/10.0] [mono][sgen] Make color visible to client permanent  (#121483)
Backport https://github.com/dotnet/runtime/pull/121247 and
https://github.com/dotnet/runtime/pull/121243 to release/10.0.
This fixes assertions like
```
* Assertion at /home/vbrezae/Xamarin/repos/runtime/src/mono/mono/metadata/sgen-tarjan-bridge.c:1151, condition `color_visible_to_client (cd)' not met
```

## Customer Impact

- [x] Customer reported
- [ ] Found internally

Some applications on maui-android can randomly crash during GC, when
using the default gc bridge (the tarjan bridge). We've had a few fixes
for the tarjan bridge merged a few months ago, but there is still this
one issue. The workaround used by customers is to fallback to an older
GC bridge which has worse performance. For some this performance impact
is not acceptable. This backport also fixes the same issue in the
CoreCLR gcbridge implementation. CoreCLR doesn't have a fallback GC
bridge implementation, so this fix is essential for the successful use
of CoreCLR/NativeAOT on android, at least for some customers.

## Regression

- [ ] Yes
- [x] No

## Testing

Tested on our own gc bridge tests, with scenario causing the issue.

## Risk

The GC bridge is a sensitive area and fixes here typically have some
carried risk. This fix however is quite straightforward, it simply pins
the value of a property inside an SCC node, rather than having it
recomputed with unstable value that was leading to problems. No changes
are done to the core algorithm. Low risk.