Reduce enabled keywords
For a basic E2E test, just enabling GC is sufficient.

Locally running this test, occasionally the test failed
because it didn't have the expected events. It's not
clear why, but with a brief tracee timeout of 1s,
collecting dotnet-trace's default keywords and enabling
all associated events is overkill for a fast E2E test.