Fix native aot outerloop (#121541)
Test merging in #121078 caused us to start compiling this test again. We
were not compiling it because more than 64k fields hits CodeView
limitations in LF_STRUCTURE/LF_CLASS (can't have more than 65534
_introduced fields_). We assert and crash compilation.

CoreCLR has it's own limitation of 65535 fields, including inherited,
but not counting `object.m_pEEType` (that doesn't exist in CoreCLR
CoreLib).

This changes the test to not have a class with > 65534 introduced
fields. This avoids the CodeView assert. We still don't generate correct
CodeView for bigger classes. Not sure if it's worth adding blocking. We
could block loading such types but 65534 is an odd number and CodeView
is an odd reason.

Cc @dotnet/ilc-contrib