Support async generic virtual methods (#121524)
* Since generic virtual method dispatch is based on
`RuntimeMethodHandle`, we need a `RuntimeMethodHandle` for async
variants. Emit the bit in the compiler, at runtime (when type loading),
and read it when needed. Reflection stack should never see this
`RuntimeMethodHandle`.
* Add `AsyncVariant` property on `MethodDesc` (since runtime type system
does this for unboxing methods too).
* Add a flag that distinguishes async variants in native layout. This is
not specific to GVMs, it also covers other cases of runtime generic
dictionary building, such as `MakeGenericMethod`.
* Add a flag that distinguishes these in generic method dictionary
hashtables (both static and built at runtime)
* Lots of piping through the extra bool