Map `CORINFO_HELP_THROWEXACT` to `THROW`
This is temporary. It will not do the right thing wrt exception stack at runtime, but it's better than crashing at compile time (which is what it does now).

Not filing a bug on this; we either fix this before we start running the tests, or hit it during test async bringup (I assume there's a test) and then ActiveIssue it as needed.