Add missing `goto` for checking encapsulation invariants.
When we encapsulate something with ML-KEM on OpenSSL, we were checking
that the amount written is what we expected, but a missing `goto` caused
it to be marked as a success.

This isn't really a reliability problem since we already know how much
encapsulation sizes are, we're just making sure OpenSSL and the managed
side agree on it.