Replace test certificate chain that uses RSA+SHA-1
The certificate chain we were using in unit tests has a root certificate
that is signed with RSA SHA-1. This signature algorithm is disabled on
some linux distributions, and it is starting to get picked up in our CI.

Let's replace the chain with one that does not use RSA+SHA-1 signatures.