JIT: Fix edge cases around async call definitions with exceptional flow
We had special logic to optimize the case where an async call defines a
local, avoiding capturing the local as part of the live state since its
value would be overridden on resumption anyway.

This does not handle the case where the callee throws an exception and
we continue internally in the function. In that case we may still have
needed to capture the local state. Consider the following example:

```csharp
int x = GetValue();
try
{
    x = await IntThrows();
}
catch
{
}

return x;
```

Here we need to capture the value of `x` in the continuation as we need
it if `IntThrows` throws. The previous logic optimized away this
capture.

For cases where we have liveness for the defined local we can use
liveness for precision when we have exceptional flow. Liveness will tell
us that `x` is live on the `IntThrows` call here, precisely because of
the exceptional flow. If there was no exceptional flow, `x` would not be
live. We can still use the previous optimization for address exposed or
tier0 cases.

Also fix when we update the live set around the async calls we are
transforming. Before this change we were updating it before the
transformation. This handling was imprecise for return buffer
definitions, but it would not usually cause problems beyond
overestimating the live set.