Add corpus support to local DotnetFuzzing runs and fuzz Deflate64 (#121019)
This PR introduces corpus support for the DotnetFuzzing project.
Currently this works for local runs only (OneFuzz requires creating a
special container for corpus, which probably needs to be done manually
-- once per fuzzer, this is left for future work).

Currently, some fuzzers (IMO incorrectly) use dictionaries as a
replacement for lack of corpus support, which makes fuzzing inefficient.
This PR prepares ground for future improvements in this regard.

To validate the concept, this PR converts ZipArchive fuzzing to use
corpus instead of a dictionary, and adds code to fuzz Deflate64 (for
which we have managed implementation internal to reading ZipArchives).