Track async variants when generating generic dictionaries (#121573)
In `GenericDictionaryNode` we have a piece of code that ensures whenever
a canonical method body of a method on a type depends on something and
we have a concrete instance of such type, the specialized thing is also
created.

For example, `class Foo<T> { Type GrabType() => typeof(List<T>); }`:
when we compile `Foo<__Canon>.GrabType` we find out we need `List<T>`.
Elsewhere we find out we need `Foo<object>`. The code in
GenericDictionaryNode that I'm modifying ensures we also generate
`List<object>`.

We do this by walking all methods on the type and generating special
`ShadowConcrete` dependencies for any canonical bodies of the methods on
the type. We were missing async variants from this reporting.

The change in EETypeNode is related, but only affects multifile
compilation that we don't care much about. I'm fixing it because it has
the same problem.