Fix check for invocation of async infrastructure methods via reflection
MethodInfo.Invoke has multiple implementation strategies. The check for async infrastructure methods needs to be placed in the code that is shared by all implementation strategies.

Before this change, this check would not kick under managed debugger or when the test is run with Switch.System.Reflection.ForceEmitInvoke=true appcontext switch.

Also, this check should fail with NotSupportedException, without wrapping it in TargetInvocationException() to match existing behaviors.