Fix collided unwind detection
A recent fix that has moved 2nd pass of the exception handling to native
code has accidentally broken detectin of collided unwind, that means
detection of when exception escapes a catch or exceptionally called finally
funclets. Unfortunately none of our coreclr and libraries test exercise
the specific case when the problem occurs, which is as follows:
* There is a try / catch inside of an outer catch
* The try block contains try / finally and the non-exceptionally called
  finally throws.
In this case, the stack walk out of the throwing finally funclet was
incorrectly classifed as an exception collision and the catch was
skipped.

The problem was that the last part of the condition to detect the collided
unwind was wrong. In fact, it was always true, so any unwind out of a
funclet was considered to be a collision even if the funclet was a
finally that was called non-exceptionally.

This change fixes it by identifying the collision by the fact that the
stack walker has moved from a funclet to native code, which is the only
case when the exception is really escaping an exceptionally called
funclet.

I have also added a regression test for the issue.

Close #121578