JIT: Fix assertion in fgGetFirstILBlock for async methods with internal conditional blocks (#121616)
The `fgGetFirstILBlock()` function was being called unconditionally during
morph to allow edge creation for tail-call-to-loop optimization.
However, this function assumes all `BBF_INTERNAL` blocks before the first
IL block are `BBJ_ALWAYS`, which is not true for async methods with
exception save/restore context where internal blocks can be conditional
(because of introducing and inlining a call to save contexts).

Fix by only calling `fgGetFirstILBlock()` when the method actually has
recursive tail calls that need to be converted to loops.