Add runtime async support to reflection dataflow analysis (#121628)
This is a port of #120683, with an extra.

The `MethodIL` that we deal with in the compiler does async variant
method signature munging. The dataflow analysis really doesn't like
that. So munge the method signatures back to how they look in metadata
and in Cecil on dataflow analysis boundaries. I don't like having to do
this, but this is a shared codebase and fewer differences is better.