MethodEntrypointOrTentativeMethod should tolerate same entrypoints for different MethodDescs (#121703)
Ran into this as I was trying something with #121456. That PR introduces
a situation where two MethodDescs could map to the same EntryPoint. We
didn't previously have that since MethodEntrypointOrTentativeMethod
doesn't support unboxing thunks (that do a similar trick).

Without this fix we would end up with two `TentativeInstanceMethodNode`
that point to the same EntryPoint and that doesn't lead to anything
good. These need to be 1:1 with entrypoints.

Cc @dotnet/ilc-contrib