Centralize virtual method invocation dependencies
`ReflectionVirtualInvokeMapNode` should be solely responsible for declaring dependencies necessary to generate the invocation table entries. Instead we had them partially duplicated between `ReflectedMethodNode` and `ReflectionVirtualInvokeMapNode.GetVirtualInvokeMapDependencies`.

This can lead to subtle problems in the unsupported `IlcTrimMetadata`==`false` mode that needs this work to happen in `ReflectionVirtualInvokeMapNode` since we don't make `ReflectedMethodNode` there (because everything is reflected).