Emit IL for AsyncResumptionStub (#121456)
Emit the expected IL for the async resumption stub in crossgen and ilc. This requires adding a MethodDesc (ExplicitContinuationAsyncMethod) that makes the hidden parameter explicit to allow passing a non-null Continuation. This follows the precedent of generic unboxing stubs in ilc. These MethodDescs are unwrapped when creating a method entrypoint node in the dependency graph.

Co-authored-by: Vladimir Sadov <vsadov@microsoft.com>
Co-authored-by: Jakob Botsch Nielsen <jakob.botsch.nielsen@gmail.com>
Co-authored-by: Michal Strehovský <MichalStrehovsky@users.noreply.github.com>